Unit test refactoring.
